Use auto-virtualenvwrapper--path as a cache for efficiency
Before this commit, auto-virtualenvwrapper-activate always invokes
auto-virtualenvwrapper-find-virtualenv-path, in order to find
virtualenv path related to the current buffer.

This might cost much because of file I/O: for example, filesystem on
slow device, deeply nested directories, many entries in upper
directories, and so on.

This commit uses auto-virtualenvwrapper--path as a cache of
auto-virtualenvwrapper-find-virtualenv-path result for efficiency.

This commit also executes caching regardless of virtualenv activation,
because caching is always effective if it is not nil.