Check validity of cached values by default-directory at caching
Before this commit, cached value auto-virtualenvwrapper--path and
auto-virtualenvwrapper--project-root are always treated as valid.

But they should be ignored, if the buffer is moved into another
directory, because the virtualenv, to which the buffer is related,
might be changed.

This commit introduces auto-virtualenvwrapper--path-cached-in and
auto-virtualenvwrapper--project-root-cached-in, in order to check
validity of cached values.

These buffer local variables hold default-directory value at caching,
and are compared with default-directory before using them.