Deactivate at switching to the buffer not related to any virtualenv
Before this commit, auto-virtualenvwrapper-activate does nothing at
switching to the buffer not related to any virtualenv. If any
virtualenv is already activated in such case, activation is (maybe,
unexpectedly) kept.

This commit makes auto-virtualenvwrapper-activate deactivate already
activated virtualenv at switching to the buffer not related to any
virtualenv.

This automatic deactivation is executed only if newly introduced
auto-virtualenvwrapper-auto-deactivate is t, and its value is nil by
default for backward compatibility.