Check if a virtualenv path is valid before setting it
This is done by checking if the path contains a `bin/python` file. (This
probably breaks on Windows, where we should check for `Scripts/python.exe`, but
I think this package already doesn't support Windows, so hopefully this isn't a
problem at the moment).

This is required e.g. when using `.python-version` for the Python version
installed with pyenv, and `.workon` for the actual venv: the current code would
use `~/.virtualenvs/3.9.13` even if that directory doesn't exist, instead of the
working venv defined by `.workon`.